pyproject.toml
[tool.poetry]
name = "invoice-processor"
version = "0.1.0"
description = "AI-powered invoice processing system with Prefect workflows"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "invoice_processor", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9,<3.14"
prefect = "^3.0.0"
pydantic = "^2.5.0"
opencv-python = "^4.8.1"
pytesseract = "^0.3.10"
PyPDF2 = "^3.0.1"
pdf2image = "^1.16.3"
pillow = "^10.1.0"
openai = "^1.6.0"
anthropic = "^0.8.0"
pandas = "^2.1.4"
python-dotenv = "^1.0.0"
typer = "^0.13.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-bdd = "^7.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
invoice-processor = "invoice_processor.main:app"
